requirements.txt
tk
opencv-python
Pillow
SpeechRecognition
pyttsx3
tkVideoPlayer
requests
python-decouple
